requirements.txt
click==8.1.7
rich==13.7.0
mysql-connector-python==8.3.0
pytest==8.1.1
pytest-mock==3.12.0
